Allow use of boilerplate file fallback also on reading configuration from file
